==> DataWork/Panel Data RSDP Impacts/02_prep_dataset_and_analysis/woreda_panel/03b_analysis_coef_each_year_results_append.R <==
# Append Individual Results

file.path(panel_rsdp_imp_data_file_path, 
          "woreda", "results_datasets", 
          "individual_datasets") %>%
  list.files(full.names = T,
             pattern = "*.Rds") %>%
  lapply(readRDS) %>%
  bind_rows() %>%
  saveRDS(file.path(panel_rsdp_imp_data_file_path, "woreda", "results_datasets",
                    "did_coef_every_year.Rds"))
